+++ src/test/java/com/wikia/webdriver/testsuites/testSuite.xml
@@ -150,7 +150,7 @@
             <class name="com.wikia.webdriver.testcases.mercurytests.old.LightboxTests"/>
             <class name="com.wikia.webdriver.testcases.auth.LoginTestsOld"/>
             <class name="com.wikia.webdriver.testcases.mercurytests.old.PortableInfoboxTests"/>
-            <class name="com.wikia.webdriver.testcases.auth.SignupTests"/>
+            <class name="com.wikia.webdriver.testcases.auth.SignupTestsOld"/>
             <class name="com.wikia.webdriver.testcases.mercurytests.old.TOCTests"/>
             <class name="com.wikia.webdriver.testcases.mercurytests.old.widgettests.AllTagsTests"/>
             <class name="com.wikia.webdriver.testcases.mercurytests.old.widgettests.ApesterTests"/>
@@ -195,7 +195,7 @@
             <class name="com.wikia.webdriver.testcases.searchtests.intrawiki.ExactMatch"/>
             <class name="com.wikia.webdriver.testcases.searchtests.intrawiki.GoSearchFeature"/>
             <class name="com.wikia.webdriver.testcases.searchtests.intrawiki.Suggestions"/>
-            <class name="com.wikia.webdriver.testcases.auth.SignUpTests"/>
+            <class name="com.wikia.webdriver.testcases.auth.SignupTests"/>
             <class name="com.wikia.webdriver.testcases.social_buttons_tests.SocialButtonsTests"/>
             <class name="com.wikia.webdriver.testcases.social.MercuryAssetCheck"/>
             <class name="com.wikia.webdriver.testcases.specialpagestests.EditAccountTests"/>
